requirements.txt
crewai>=0.32.0
crewai-tools>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.22
sentence-transformers>=2.3.0
PyPDF2>=3.0.0
python-docx>=1.1.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
